+++ package.json
@@ -101,17 +101,10 @@
     "not op_mini all"
   ],
   "dependencies": {
-<<<<<<< HEAD
     "@accordproject/markdown-html": "^0.10.4",
     "@accordproject/markdown-slate": "^0.10.4",
     "@babel/runtime": "^7.8.7",
     "css-loader": "^3.4.2",
-=======
-    "@accordproject/markdown-html": "0.10.2-20200305210058",
-    "@accordproject/markdown-slate": "0.10.2-20200305210058",
-    "@babel/runtime": "^7.8.4",
-    "css-loader": "^3.2.0",
->>>>>>> Critical fix to dependencies for proper markdown transform version
     "is-hotkey": "^0.1.6",
     "prop-types": "^15.7.2",
     "react-textarea-autosize": "^7.1.2",
